chore(code-splitting): refactor folder structure
