+++ packages/dotcom-build-code-splitting/src/__test__/__fixtures__/@financial-times/privacy-module-1/package.json
