+++ packages/dotcom-build-code-splitting/src/__test__/__fixtures__/entry-point.js
